removed spacing and shape attribute from Cubic, moved to topotools, updated tests
